fix(macos): resolve C compilation warnings and Tcl/Tk linking for standalone build
- Patched C code (pointer casts, unused/uninitialized vars, mismatched signatures, FREAD/FWRITE macros) in memops/global and ccpnmr/analysis.
- Updated setup.py to link against Homebrew's tcl9tk9.0 and tcl9.0 instead of tk8.6.
- Modified setup.py to export fully-qualified extension module names (e.g. memops.c.StoreFile) to fix runtime import failures.
- Patched make-standalone-macos.sh to enforce PYTHONHOME during pip install and use --force-reinstall.

+++ ccpnmr2.5/c/memops/global/hash_table.c
@@ -76,7 +76,7 @@ Bool equal_pointers(Hash_key key1, Hash_key key2)
 
 Hash_value hash_pointers(Hash_key key)
 {
-    return create_hash((int) key);
+    return create_hash((int)(size_t) key);
 }
 
 static Hash_entry *next_entry(Hash_table hash_table, Hash_entry *entry)
